4.20.19
 - Even more clang tidy-fix changes. Only changes left now are even more c-style casting changes, and magic number fixes.

+++ src/Modules/Graphics/Lighting/Reflector/ReflectorScheduler_System.cpp
@@ -88,6 +88,6 @@ void ReflectorScheduler_System::updateComponents(const float& /*deltaTime*/, con
 
 		// Resize the reflector map to fit number of entities this frame
 		m_frameData.envmapFBO.resize(m_frameData.envmapSize, static_cast<unsigned int>(cameraCount));
-		m_frameData.reflectorLayers = cameraCount;
+		m_frameData.reflectorLayers = static_cast<size_t>(cameraCount);
 	}
 }
